tool/verif: split mods & avoid global symbex
